O ID, agora, está gerando automaticamente, a busca ainda não foi concluida persiste o erro que dá quando aperta o button de pesquisa. Agora as figuras que foram implementadas na revisão passada quando acessa uma página e volta elas não aparecem mais, tem que revisar isso e também a aprte de edição. No mais está faltando só o layout e a busca que é apenas do protocolo e do processo.
The ID now is generating automatically, the search has not been completed, there remains an error that occurs when pressing the search button. Now the figures have been implemented in the review month when accessing a page back and they appear no more, has to review it and also aprte editing. No more missing only the layout and search that is just the protocol and procedure.